complete some error for this app, add session for role, add feature reset password for user

+++ skill-data.php
@@ -57,7 +57,7 @@ $firstData = ( $amountData1Page * $activePage ) - $amountData1Page;
         <ul class="navbar-nav navbar-right">
           <li class="dropdown"><a href="#" data-toggle="dropdown" class="nav-link dropdown-toggle nav-link-lg nav-link-user">
             <img alt="image" src="assets/img/avatar/avatar-1.png" class="rounded-circle mr-1">
-            <div class="d-sm-none d-lg-inline-block">Hi, <?= $_SESSION['login']; ?></div></a>
+            <div class="d-sm-none d-lg-inline-block">Hi, <?= $_SESSION['login']['name']; ?></div></a>
             <div class="dropdown-menu dropdown-menu-right">
               <a href="logout.php" class="dropdown-item has-icon text-danger">
                 <i class="fas fa-sign-out-alt"></i> Logout
@@ -115,7 +115,7 @@ $firstData = ( $amountData1Page * $activePage ) - $amountData1Page;
                       <td><?= $row['skill'] ?></td>
                       <td>
                         <a href="features-profile-skill.php?id=<?= $row['sid'] ?>" class="btn btn-success mr-1">Edit</a>
-                        <a href="delete-skill.php?id=<?= $row['sid'] ?>" class="btn btn-danger">Delete</a>
+                        <a href="delete-skill.php?id=<?= $row['sid'] ?>" class="btn btn-danger" onclick="return confirm('yakin ingin menghapus?')">Delete</a>
                       </td>
                     </tr>
                   <?php $i++; } ?>
